+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `ping()` method
+- `readTimeout` and `writeTimeout` settings for operations timeout support
+
 ## [1.1.0] - 2020-03-10
 ## Added
 - `getSchema` method